Lecture 123 - [ExecuteInEditMode] - Done.
Most of this lecture was spent struggling to understand Ben's explanation ofthe snap to grid script that we made. Throughout I was confused.

I learned that I don't remember much about Range arrays and that Unity docs for it almost don't exsist.

I learned that we can run scripts in the editor, which are useful for controlling what the designer can and can't do. We use this to control the snapping to grid ofhte cubes.