Lecture 124 - Using Text Mesh For Labels - Done.
We learned about text meshs, font size, character size and resolution.
Also an interesting bit of code that I don't remember what it is but I remember what it does and that it's near the top of the script.
It makes it so that you click on the parent object in the scene window rather than the child objects. This stops the designer breaking things.