Lecture 131 - Finding Game Objects By Name - Done.
It actually turned into more than that as we had to use a serialize field of two waypoints, named start and end. There was a challenge that I failed because Ben threw me off by talking about the wrong script as an option. It was a bit of a pain and it's taken me all day to wake up, let alone do anything useful.