Revise Coroutines & Much More - Done
I gave up on the challenge but I did make some progress by getting the enemy to spawn once but I needed it to do that continously. Anyway Ben wants to cover spawning in the next lecture as he got the coroutine working and printing spawning forvever but now the challenge is having it spawn under the right parent object and at the start location and so on.