Lecture 147 - Target Closest Enemy -
Apparently all he had to do was change it to 4 rather than 2 seconds before spawn and it stopped looking like it was only spawning right behind it.

So my code would have worked if the while true was added. The range thing and using seconds between spawns instead of Vector3 0,0,0 would have made it cleaner. That's all that had to be fixed.

Anyway then Ben got the tower to check for the closest enemy. I failed the main challenge and didn't even really notice him set a mini challenge. I don't seem to make use or var that much and I'm still confused by almost every way that they explain what needs to be done in a challenge. So fuck it all.