Lecture 149 - Conditional Instantiation - Started
I did half of it but I just stopped caring. I may continue tonight or tomorrow.

What we did do was get the mouse clicking on the blocks and printing only when a click on a block happens. Also we were working through getting it to only do that for blocks that it should do that for. I still don't know why aren't letting them be placed on blocks that are on the path.