Lecture 149: Conditional Instantiation - Done
It's done. It was all straight forward except for errors that Ben made that I spotted one of and another person spotted it in the q and a. The clicks now check if it's a placeable block, even if I still need to fully understand some C# stuff like this example:

Say you have void Method(method) what is the small method. What is the thing that it's called.