Lecture 151 - Particles & Algorithm Improvements - Started/Broken
Ben fixes an error that I noticed two lectures ago. Then we transfer in the Enemy prefab and I notice that now it spawns/moves to 7,5 and then moves all over the place. Also the towers don't shoot and then Enemies don't child themselves to the prefab. I tried using an earlier Enemy prefab but that made no difference.