Lecture 151 - Particles & Algorithm Improvements -
I checked a copy of Ben's pathfinder script and found that I had SetAsPath(endWaypoint) instead of SetAsPath(startWaypoint) in the CreatePath method.
Why is it always such small things that I miss! Well at least that fixed it.
The basic idea is the next line of path.Reverse was going to endWaypoint instead of startWaypoint, which is why I got 7,5 instead of 0,0.

I failed to fix the Enemies not becoming children issue or the towers not shooting.