Lecture 153 - Implementing A Ring Buffer - Done
We added some of the Tower factory script and we moved some of the waypoint script into it and I did most of achallenge. I didn't get it right of course but I was kind of in the right area, just overly complicated as per usual. Anyway let's just move on I really can't bear this course section much longer.