Lecture 154 - Revising C# Queues -
We added code to make a queue that adds the towers to it each time it instantiates.
It de-queues the towers when it reaches it's limit.

In the next lecture we are doing something waypoints that I don't understand and is placeable that I think I understand and soon will be proven wrong as well as destorying particle clones.